minor fixes for api
