+++ utils/limiter.js
@@ -1,0 +1,12 @@
+const rateLimit = require('express-rate-limit');
+
+const limiter = rateLimit({
+  windowMs: 10 * 60 * 1000,
+  max: 100,
+  standardHeaders: true,
+  legacyHeaders: false,
+});
+
+module.exports = {
+  limiter,
+};
