+++ utils/validation.js
@@ -25,6 +25,7 @@ const userIdSchemaValidate = celebrate({
 const profileSchemaValidate = celebrate({
   body: Joi.object().keys({
     name: Joi.string().min(2).max(30),
+    email: Joi.string().required().email(),
   }),
 });
